requirements.txt
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.2
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
easydict==1.7
entrypoints==0.3
filterpy==1.4.4
gunicorn==20.0.4
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.13.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
llvmlite==0.31.0
MarkupSafe==1.1.1
mistune==0.8.4
munkres==1.0.12
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
panda==0.3.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
progressbar2==3.38.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
python-utils==2.4.0
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
ruamel.yaml==0.15.81
scikit-learn==0.19.2
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Unipath==1.1
utm==0.4.2
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
xlrd==1.1.0
zipp==3.1.0
